refactor(explore): unify data layer into useExploreData hook
explore.tsx (native) and explore.web.tsx (web) each carried their own
copy of the Home data layer — ~28 fetches across a mount effect + a
per-user effect — that had drifted in limits, field names, and one
screen's dead code. Extract it all into a new platform-neutral hook,
src/hooks/useExploreData.ts, returning one flat snapshot. Each view
keeps its own presentation: native its virtualised FlatList + `rows`
memo, web its grid.

Reconciled to the more mature option from each side:
- Limits: iconic 25, horror 25, eras 8, spotlight pool 10 (web's). The
  hook returns the un-sliced spotlight pool; native slices to 5, web
  sizes responsively.
- heroCount: `number | null` (null = loading); native coerces `?? 0`.
- matchup: `TodaysMatchup | null | undefined` tri-state so web keeps its
  loading skeleton; native's `if (matchup)` guard is unchanged.
- Unified the gate flag (native `initialLoaded` / web `homeStarted`) to
  `started`, and the `franchise`/`franchiseIcons` name; views alias on
  destructure so their render trees are untouched.
- Dropped the web-only strongestHero/smartestHero/fastestHero/
  publisherCounts fetches — they were fetched but never rendered (4 fewer
  queries per web load).

Net: native explore.tsx 678->532, web explore.web.tsx 1656->1456; the
data layer now lives once in a 255-line hook. typecheck clean; format
clean; 363 tests pass; 0 lint errors.

Behavior to verify in-app (web + native): row contents (a few more items
on native's iconic/horror/eras), spotlight billboard, matchup skeleton.
